Refactoring domain service
