+++ backend/src/main/java/rs/in/andric/demo/PraksaPrvi/service/DomainService.java
@@ -1,59 +1,25 @@
 package rs.in.andric.demo.PraksaPrvi.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import rs.in.andric.demo.PraksaPrvi.model.Domain;
-import rs.in.andric.demo.PraksaPrvi.model.News;
-import rs.in.andric.demo.PraksaPrvi.repository.DomainRepository;
 
 import java.util.List;
-import java.util.Optional;
 
-@Service
-public class DomainService {
 
-    @Autowired
-    private DomainRepository domainRepository;
+public interface DomainService {
 
-    public List<Domain> getAll(){
-        return domainRepository.findAll();
-    }
 
-    public Domain getDomainById(Integer id){
-        Optional<Domain> domain = domainRepository.findById(id);
-        if(!domain.isPresent()){
-            return new Domain();
-        }
-        return domain.get();
-    }
+    List<Domain> getAll();
 
-    public boolean hasDomainWithName(String name){
-        Optional<Domain> domain = domainRepository.findByName(name);
-        if(!domain.isPresent()){
-            return false;
-        }
-        return true;
-    }
+    Domain getDomainById(Integer id);
 
-    public Domain getDomainByName(String name){
-        Optional<Domain> domain = domainRepository.findByName(name);
-        if(!domain.isPresent()){
-            return new Domain();
-        }
-        return domain.get();
-    }
+    boolean hasDomainWithName(String name);
 
-    public Domain addDomain(Domain domain){
-        return domainRepository.save(domain);
-    }
+    Domain getDomainByName(String name);
 
-    public Domain updateDomain(Integer id, Domain domain){
-        domain.setId(id);
-        return domainRepository.save(domain);
-    }
+    Domain addDomain(Domain domain);
 
-    public void deleteDomain(Integer id){
-        domainRepository.deleteById(id);
-    }
+    Domain updateDomain(Integer id, Domain domain);
+
+    void deleteDomain(Integer id);
 
 }
