+++ backend/src/main/java/rs/in/andric/demo/PraksaPrvi/service/impl/DomainServiceImpl.java
@@ -1,0 +1,61 @@
+package rs.in.andric.demo.PraksaPrvi.service.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import rs.in.andric.demo.PraksaPrvi.model.Domain;
+import rs.in.andric.demo.PraksaPrvi.repository.DomainRepository;
+import rs.in.andric.demo.PraksaPrvi.service.DomainService;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class DomainServiceImpl implements DomainService {
+
+	@Autowired
+	private DomainRepository domainRepository;
+
+
+	public List<Domain> getAll(){
+		return domainRepository.findAll();
+	}
+
+	public Domain getDomainById(Integer id){
+		Optional<Domain> domain = domainRepository.findById(id);
+		if(!domain.isPresent()){
+			return new Domain();
+		}
+		return domain.get();
+	}
+
+	public boolean hasDomainWithName(String name){
+		Optional<Domain> domain = domainRepository.findByName(name);
+		if(!domain.isPresent()){
+			return false;
+		}
+		return true;
+	}
+
+	public Domain getDomainByName(String name){
+		Optional<Domain> domain = domainRepository.findByName(name);
+		if(!domain.isPresent()){
+			return new Domain();
+		}
+		return domain.get();
+	}
+
+	public Domain addDomain(Domain domain){
+		return domainRepository.save(domain);
+	}
+
+	public Domain updateDomain(Integer id, Domain domain){
+		domain.setId(id);
+		return domainRepository.save(domain);
+	}
+
+	public void deleteDomain(Integer id){
+		domainRepository.deleteById(id);
+	}
+
+
+}
